Crate common_procedure

Common traits and structures for the procedure framework.

Re-exports

pub use crate::error::Error;

pub use crate::error::Result;

pub use crate::watcher::Watcher;

Modules

error

local

options

Common traits and structures for the procedure framework.

procedure 🔒

store

watcher

Structs

Context

Procedure execution context.

LockKey

Keys to identify required locks.

ParseIdError

ProcedureId

Unique id for Procedure.

ProcedureWithId

A procedure with specific id.

Enums

ProcedureState

State of a submitted procedure.

Status

Procedure execution status.

StringKey

Traits

ContextProvider

ContextProvider provides information about procedures in the ProcedureManager.

Procedure

A Procedure represents an operation or a set of operations to be performed step-by-step.

ProcedureManager

ProcedureManager executes Procedure submitted to it.

Type Aliases

BoxedProcedure

Boxed Procedure.

BoxedProcedureLoader

Loader to recover the Procedure instance from serialized data.

Output

ProcedureManagerRef

Ref-counted pointer to the ProcedureManager.