Enum common_procedure::Status

pub enum Status {
    Executing {
        persist: bool,
    },
    Suspended {
        subprocedures: Vec<ProcedureWithId>,
        persist: bool,
    },
    Done,
}

Procedure execution status.

Variants

Executing

Fields

persist: bool

Whether the framework needs to persist the procedure.

The procedure is still executing.

Suspended

Fields

subprocedures: Vec<ProcedureWithId>

persist: bool

Whether the framework needs to persist the procedure.

The procedure has suspended itself and is waiting for subprocedures.

Done

the procedure is done.

Implementations

impl Status

pub fn executing(persist: bool) -> Status

Returns a Status::Executing with given persist flag.

pub fn need_persist(&self) -> bool

Returns true if the procedure needs the framework to persist its intermediate state.