Enum common_procedure::error::Error

pub enum Error {

    External {
        source: BoxedError,
    },
    LoaderConflict {
        name: String,
        location: Location,
    },
    ManagerNotStart {
        location: Location,
    },
    ToJson {
        error: Error,
        location: Location,
    },
    DuplicateProcedure {
        procedure_id: ProcedureId,
        location: Location,
    },
    PutState {
        key: String,
        location: Location,
        source: BoxedError,
    },
    DeleteState {
        key: String,
        error: Error,
    },
    DeleteStates {
        keys: String,
        location: Location,
        source: BoxedError,
    },
    ListState {
        path: String,
        location: Location,
        source: BoxedError,
    },
    FromJson {
        error: Error,
        location: Location,
    },
    RetryLater {
        source: BoxedError,
    },
    ProcedurePanic {
        procedure_id: ProcedureId,
    },
    WaitWatcher {
        error: RecvError,
        location: Location,
    },
    ProcedureExec {
        source: Arc<Error>,
        location: Location,
    },
    RetryTimesExceeded {
        source: Arc<Error>,
        procedure_id: ProcedureId,
    },
    CorruptedData {
        error: FromUtf8Error,
    },
    StartRemoveOutdatedMetaTask {
        source: Error,
        location: Location,
    },
    StopRemoveOutdatedMetaTask {
        source: Error,
        location: Location,
    },
    SubprocedureFailed {
        subprocedure_id: ProcedureId,
        source: Arc<Error>,
        location: Location,
    },

}

Procedure error.

Variants

External

Fields

source: BoxedError

LoaderConflict

Fields

name: String

location: Location

ManagerNotStart

Fields

location: Location

ToJson

Fields

error: Error

location: Location

DuplicateProcedure

Fields

procedure_id: ProcedureId

location: Location

PutState

Fields

key: String

location: Location

source: BoxedError

DeleteState

Fields

key: String

error: Error

DeleteStates

Fields

keys: String

location: Location

source: BoxedError

ListState

Fields

path: String

location: Location

source: BoxedError

FromJson

Fields

error: Error

location: Location

RetryLater

Fields

source: BoxedError

ProcedurePanic

Fields

procedure_id: ProcedureId

WaitWatcher

Fields

error: RecvError

location: Location

ProcedureExec

Fields

source: Arc<Error>

location: Location

RetryTimesExceeded

Fields

source: Arc<Error>

procedure_id: ProcedureId

CorruptedData

Fields

error: FromUtf8Error

StartRemoveOutdatedMetaTask

Fields

source: Error

location: Location

StopRemoveOutdatedMetaTask

Fields

source: Error

location: Location

SubprocedureFailed

Fields

subprocedure_id: ProcedureId

source: Arc<Error>

location: Location

Implementations

impl Error

pub fn external<E: ErrorExt + Send + Sync + 'static>(err: E) -> Error

Creates a new Error::External error from source err.

pub fn retry_later<E: ErrorExt + Send + Sync + 'static>(err: E) -> Error

Creates a new Error::RetryLater error from source err.

pub fn is_retry_later(&self) -> bool

Determine whether it is a retry later type through StatusCode

pub fn from_error_ext<E: ErrorExt + Send + Sync + 'static>(err: E) -> Self

Creates a new Error::RetryLater or Error::External error from source err according to its StatusCode.