common_procedure::local

Type Alias DynamicKeyLock

pub(crate) type DynamicKeyLock = Arc<KeyRwLock<String>>;

The dynamic lock for procedure execution.

Unlike the procedure-level locks, these locks are acquired dynamically by the procedure during execution. They are only held when the procedure specifically needs these keys and are released as soon as the procedure no longer needs them. This allows for more fine-grained concurrency control during procedure execution.

Aliased Type

pub(crate) struct DynamicKeyLock { /* private fields */ }