common_procedure::store
Module util

Structs

CollectingState 🔒

Functions

merge_multiple_values 🔒

Merges multiple values into a single key-value pair. Returns an error if:

multiple_value_stream

Merges multiple values that have the same prefix of the key from upstream into a single value.

parse_segments 🔒

Type Aliases

Upstream 🔒