Trait common_procedure::Procedure

pub trait Procedure: Send + Sync {
    // Required methods
    fn type_name(&self) -> &str;
    fn execute<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ctx: &'life1 Context
    ) -> Pin<Box<dyn Future<Output = Result<Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn dump(&self) -> Result<String>;
    fn lock_key(&self) -> LockKey;
}

A Procedure represents an operation or a set of operations to be performed step-by-step.

Required Methods

fn type_name(&self) -> &str

Type name of the procedure.

fn execute<'life0, 'life1, 'async_trait>( &'life0 mut self, ctx: &'life1 Context ) -> Pin<Box<dyn Future<Output = Result<Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Execute the procedure.

The implementation must be idempotent.

fn dump(&self) -> Result<String>

Dump the state of the procedure to a string.

fn lock_key(&self) -> LockKey

Returns the LockKey that this procedure needs to acquire.