common_procedure

Enum Status

pub enum Status {
    Executing {
        persist: bool,
        clean_poisons: bool,
    },
    Suspended {
        subprocedures: Vec<ProcedureWithId>,
        persist: bool,
    },
    Poisoned {
        keys: PoisonKeys,
        error: Error,
    },
    Done {
        output: Option<Output>,
    },
}

Procedure execution status.

Variants

Executing

The procedure is still executing.

Fields

persist: bool

Whether the framework needs to persist the procedure.

clean_poisons: bool

Whether the framework needs to clean the poisons.

Suspended

The procedure has suspended itself and is waiting for subprocedures.

Fields

subprocedures: Vec<ProcedureWithId>

persist: bool

Whether the framework needs to persist the procedure.

Poisoned

The procedure is poisoned.

Fields

keys: PoisonKeys

The keys that cause the procedure to be poisoned.

error: Error

The error that cause the procedure to be poisoned.

Done

the procedure is done.

Fields

output: Option<Output>

Implementations

impl Status

pub fn poisoned( keys: impl IntoIterator<Item = PoisonKey>, error: Error, ) -> Status

Returns a Status::Poisoned with given keys and error.

pub fn executing(persist: bool) -> Status

Returns a Status::Executing with given persist flag.

pub fn executing_with_clean_poisons(persist: bool) -> Status

Returns a Status::Executing with given persist flag and clean poisons.

pub fn done() -> Status

Returns a Status::Done without output.

pub fn downcast_output_ref<T: 'static>(&self) -> Option<&T>

Downcasts Status::Done’s output to &T #Panic:

if Status is not the Status::Done.

if the output is None.

pub fn done_with_output<T: Any + Send + Sync>(output: T) -> Status

Returns a Status::Done with output.

pub fn is_done(&self) -> bool

Returns true if the procedure is done.

pub fn need_persist(&self) -> bool

Returns true if the procedure needs the framework to persist its intermediate state.

pub fn need_clean_poisons(&self) -> bool

Returns true if the framework needs to clean the poisons.