common_procedure::error

Enum Error

pub enum Error {

    External {
        source: BoxedError,
        clean_poisons: bool,
    },
    LoaderConflict {
        name: String,
        location: Location,
    },
    ManagerNotStart {
        location: Location,
    },
    ToJson {
        error: Error,
        location: Location,
    },
    DuplicateProcedure {
        procedure_id: ProcedureId,
        location: Location,
    },
    TooManyRunningProcedures {
        max_running_procedures: usize,
        location: Location,
    },
    PutState {
        key: String,
        location: Location,
        source: BoxedError,
    },
    PutPoison {
        key: String,
        token: String,
        location: Location,
        source: BoxedError,
    },
    GetPoison {
        key: String,
        location: Location,
        source: BoxedError,
    },
    DeletePoison {
        key: String,
        token: String,
        location: Location,
        source: BoxedError,
    },
    DeleteState {
        key: String,
        error: Error,
    },
    DeleteStates {
        keys: String,
        location: Location,
        source: BoxedError,
    },
    ListState {
        path: String,
        location: Location,
        source: BoxedError,
    },
    FromJson {
        error: Error,
        location: Location,
    },
    RetryLater {
        source: BoxedError,
        clean_poisons: bool,
    },
    ProcedurePanic {
        procedure_id: ProcedureId,
    },
    WaitWatcher {
        error: RecvError,
        location: Location,
    },
    ProcedureExec {
        source: Arc<Error>,
        location: Location,
    },
    RollbackProcedureRecovered {
        error: String,
        location: Location,
    },
    RetryTimesExceeded {
        source: Arc<Error>,
        procedure_id: ProcedureId,
    },
    RollbackTimesExceeded {
        source: Arc<Error>,
        procedure_id: ProcedureId,
    },
    StartRemoveOutdatedMetaTask {
        source: Error,
        location: Location,
    },
    StopRemoveOutdatedMetaTask {
        source: Error,
        location: Location,
    },
    ParseSegmentKey {
        location: Location,
        key: String,
        error: ParseIntError,
    },
    Unexpected {
        location: Location,
        err_msg: String,
    },
    RollbackNotSupported {
        location: Location,
    },
    ProcedureNotFound {
        procedure_id: ProcedureId,
        location: Location,
    },
    PoisonKeyNotDefined {
        key: PoisonKey,
        procedure_id: ProcedureId,
        location: Location,
    },

}

Procedure error.

Variants

External

Fields

source: BoxedError

clean_poisons: bool

LoaderConflict

Fields

name: String

location: Location

ManagerNotStart

Fields

location: Location

ToJson

Fields

error: Error

location: Location

DuplicateProcedure

Fields

procedure_id: ProcedureId

location: Location

TooManyRunningProcedures

Fields

max_running_procedures: usize

location: Location

PutState

Fields

key: String

location: Location

source: BoxedError

PutPoison

Fields

key: String

token: String

location: Location

source: BoxedError

GetPoison

Fields

key: String

location: Location

source: BoxedError

DeletePoison

Fields

key: String

token: String

location: Location

source: BoxedError

DeleteState

Fields

key: String

error: Error

DeleteStates

Fields

keys: String

location: Location

source: BoxedError

ListState

Fields

path: String

location: Location

source: BoxedError

FromJson

Fields

error: Error

location: Location

RetryLater

Fields

source: BoxedError

clean_poisons: bool

ProcedurePanic

Fields

procedure_id: ProcedureId

WaitWatcher

Fields

error: RecvError

location: Location

ProcedureExec

Fields

source: Arc<Error>

location: Location

RollbackProcedureRecovered

Fields

error: String

location: Location

RetryTimesExceeded

Fields

source: Arc<Error>

procedure_id: ProcedureId

RollbackTimesExceeded

Fields

source: Arc<Error>

procedure_id: ProcedureId

StartRemoveOutdatedMetaTask

Fields

source: Error

location: Location

StopRemoveOutdatedMetaTask

Fields

source: Error

location: Location

ParseSegmentKey

Fields

location: Location

key: String

error: ParseIntError

Unexpected

Fields

location: Location

err_msg: String

RollbackNotSupported

Fields

location: Location

ProcedureNotFound

Fields

procedure_id: ProcedureId

location: Location

PoisonKeyNotDefined

Fields

key: PoisonKey

procedure_id: ProcedureId

location: Location

Implementations

impl Error

pub fn external<E: ErrorExt + Send + Sync + 'static>(err: E) -> Error

Creates a new Error::External error from source err.

pub fn external_and_clean_poisons<E: ErrorExt + Send + Sync + 'static>( err: E, ) -> Error

Creates a new Error::External error from source err and clean poisons.

pub fn retry_later<E: ErrorExt + Send + Sync + 'static>(err: E) -> Error

Creates a new Error::RetryLater error from source err.

pub fn retry_later_and_clean_poisons<E: ErrorExt + Send + Sync + 'static>( err: E, ) -> Error

Creates a new Error::RetryLater error from source err and clean poisons.

pub fn is_retry_later(&self) -> bool

Determine whether it is a retry later type through [StatusCode]

pub fn need_clean_poisons(&self) -> bool

Determine whether it needs to clean poisons.

pub fn from_error_ext<E: ErrorExt + Send + Sync + 'static>(err: E) -> Self

Creates a new Error::RetryLater or Error::External error from source err according to its [StatusCode].