Struct common_procedure::local::ManagerContext

pub(crate) struct ManagerContext {
    loaders: Mutex<HashMap<String, Box<dyn Fn(&str) -> Result<BoxedProcedure> + Send>>>,
    lock_map: LockMap,
    procedures: RwLock<HashMap<ProcedureId, Arc<ProcedureMeta>>>,
    messages: Mutex<HashMap<ProcedureId, ProcedureMessage>>,
    finished_procedures: Mutex<VecDeque<(ProcedureId, Instant)>>,
}

Shared context of the manager.

Fields
loaders: Mutex<HashMap<String, Box<dyn Fn(&str) -> Result<BoxedProcedure> + Send>>>

Procedure loaders. The key is the type name of the procedure which the loader returns.

lock_map: LockMapprocedures: RwLock<HashMap<ProcedureId, Arc<ProcedureMeta>>>messages: Mutex<HashMap<ProcedureId, ProcedureMessage>>

Messages loaded from the procedure store.

finished_procedures: Mutex<VecDeque<(ProcedureId, Instant)>>

Ids and finished time of finished procedures.

Implementations

impl ManagerContext

fn new() -> ManagerContext

Returns a new ManagerContext.

fn contains_procedure(&self, procedure_id: ProcedureId) -> bool

Returns true if the procedure with specific procedure_id exists.

fn try_insert_procedure(&self, meta: Arc<ProcedureMeta>) -> bool

Try to insert the procedure to the context if there is no procedure with same ProcedureId.

Returns false if there is already a procedure using the same ProcedureId.

fn state(&self, procedure_id: ProcedureId) -> Option<ProcedureState>

Returns the ProcedureState of specific procedure_id.

fn watcher(&self, procedure_id: ProcedureId) -> Option<Watcher>

Returns the Watcher of specific procedure_id.

fn notify_by_subprocedure(&self, procedure_id: ProcedureId)

Notify a suspended parent procedure with specific procedure_id by its subprocedure.

fn load_one_procedure( &self, procedure_id: ProcedureId ) -> Option<LoadedProcedure>

Load procedure with specific procedure_id from cached ProcedureMessages.

fn load_one_procedure_from_message( &self, procedure_id: ProcedureId, message: &ProcedureMessage ) -> Option<LoadedProcedure>

Load procedure from specific ProcedureMessage.

fn procedures_in_tree(&self, root: &Arc<ProcedureMeta>) -> Vec<ProcedureId>

Returns all procedures in the tree (including given root procedure).

If callers need a consistent view of the tree, they must ensure no new procedure is added to the tree during using this method.

fn find_procedures( &self, procedure_ids: &[ProcedureId], metas: &mut Vec<Arc<ProcedureMeta>> )

Finds procedures by given procedure_ids.

Ignores the id if corresponding procedure is not found.

fn remove_messages(&self, procedure_ids: &[ProcedureId])

Remove cached ProcedureMessage by ids.

fn on_procedures_finish(&self, procedure_ids: &[ProcedureId])

Clean resources of finished procedures.

fn remove_outdated_meta(&self, ttl: Duration)

Remove metadata of outdated procedures.