common_procedure

Module procedure

Structs

Context

Procedure execution context.

LockKey

Keys to identify required locks.

ParseIdError

ParseIdSnafu ๐

SNAFU context selector for the ParseIdError error

PoisonKey

PoisonKeys

A collection of PoisonKeys.

ProcedureId

Unique id for Procedure.

ProcedureInfo

ProcedureWithId

A procedure with specific id.

Enums

InitProcedureState

The initial procedure state.

ProcedureState

State of a submitted procedure.

Status

Procedure execution status.

StringKey

Traits

ContextProvider

ContextProvider provides information about procedures in the ProcedureManager.

Procedure

A Procedure represents an operation or a set of operations to be performed step-by-step.

ProcedureManager

ProcedureManager executes Procedure submitted to it.

Type Aliases

BoxedProcedure

Boxed Procedure.

BoxedProcedureLoader

Loader to recover the Procedure instance from serialized data.

ContextProviderRef

Reference-counted pointer to ContextProvider.

Output

ProcedureManagerRef

Ref-counted pointer to the ProcedureManager.